admin/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-frontmatter==1.0.1
markdown==3.5.1
Werkzeug==3.0.1
